# generated/swagger.json
{
	"components": {
...
		"schemas": {
			"SelectValue": {
				"properties": {
					"type": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"value",
					"id"
				],
				"type": "object"
			},
...
			"GoalSummery": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cycle": {
						"$ref": "#/components/schemas/PerformanceCycle"
					},
					"cycleId": {
						"type": "number",
						"format": "double"
					},
					"owner": {
						"$ref": "#/components/schemas/GoalHolder"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"goals": {
						"items": {
							"$ref": "#/components/schemas/Goal"
						},
						"type": "array"
					},
					"approvalStatus": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"completed": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"duration": {
						"items": {
							"$ref": "#/components/schemas/TimeDuration"
						},
						"type": "array"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"teamId": {
						"type": "number",
						"format": "double"
					},
					"teamLeader": {
						"type": "boolean"
					},
					"peerReview": {
						"items": {
							"$ref": "#/components/schemas/PeerReview"
						},
						"type": "array"
					},
					"subordinateReviews": {
						"items": {
							"$ref": "#/components/schemas/SubordinateReview"
						},
						"type": "array"
					},
					"criteriaSet": {
						"$ref": "#/components/schemas/CriteriaSet"
					},
					"criteriaSetId": {
						"type": "number",
						"format": "double"
					},
					"configurations": {
						"items": {
							"$ref": "#/components/schemas/TeamConfiguration"
						},
						"type": "array"
					},
					"leadConfigurations": {
						"items": {
							"$ref": "#/components/schemas/TeamConfiguration"
						},
						"type": "array"
					},
					"internalCustomers": {
						"items": {
							"$ref": "#/components/schemas/InternalCustomer"
						},
						"type": "array"
					},
					"comment": {
						"$ref": "#/components/schemas/SupervisorAnalysis"
					},
					"internalCustomerApprover": {
						"type": "number",
						"format": "double"
					},
					"publishForPeers": {
						"type": "boolean"
					},
					"publishForLeaderPeers": {
						"type": "boolean"
					},
					"publishForInternalCustomers": {
						"type": "boolean"
					},
					"cycleClose": {
						"type": "boolean"
					},
					"finalCommentSubmit": {
						"type": "boolean"
					},
					"approvedInternalCustomers": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerMarks": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerLeaderMarks": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerCount": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerWeight": {
						"type": "number",
						"format": "double"
					},
					"internalCustomerLeaderWeight": {
						"type": "number",
						"format": "double"
					},
					"approvedPeerReview": {
						"type": "number",
						"format": "double"
					},
					"peerReviewMarks": {
						"type": "number",
						"format": "double"
					},
					"peerReviewCount": {
						"type": "number",
						"format": "double"
					},
					"peerReviewWeight": {
						"type": "number",
						"format": "double"
					},
					"teamPerformanceWeight": {
						"type": "number",
						"format": "double"
					},
					"approvedTeamPerformance": {
						"type": "number",
						"format": "double"
					},
					"teamPerformanceCount": {
						"type": "number",
						"format": "double"
					},
					"teamPerformanceMarks": {
						"type": "number",
						"format": "double"
					},
					"supervisorFeedBackWeight": {
						"type": "number",
						"format": "double"
					},
					"approvedSupervisorFeedBack": {
						"type": "number",
						"format": "double"
					},
					"supervisorFeedBackCount": {
						"type": "number",
						"format": "double"
					},
					"supervisorFeedBackMarks": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"cycle",
					"owner",
					"goalsAvailability",
					"goals",
					"duration",
					"team",
					"peerReview",
					"subordinateReviews",
					"criteriaSet",
					"configurations",
					"leadConfigurations",
					"internalCustomers",
					"comment",
					"cycleClose",
					"finalCommentSubmit"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Team": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"teamType": {
						"$ref": "#/components/schemas/TeamType"
					},
					"teamTypeId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"leads": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"parent": {
						"$ref": "#/components/schemas/Team"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"userGroupMembers": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					},
					"userGroupMembers1": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					},
					"image": {
						"type": "string"
					},
					"goalHolder": {
						"$ref": "#/components/schemas/GoalHolder"
					},
					"goalHolderId": {
						"type": "number",
						"format": "double"
					},
					"memberCount": {
						"type": "number",
						"format": "double"
					},
					"cycle": {
						"$ref": "#/components/schemas/PerformanceCycle"
					},
					"cycleId": {
						"type": "number",
						"format": "double"
					},
					"previous": {
						"$ref": "#/components/schemas/Team"
					},
					"previousId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"teamType",
					"leads",
					"teams",
					"children",
					"members",
					"goalHolder",
					"memberCount",
					"cycle",
					"previous"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"employeeId": {
						"type": "string"
					},
					"epfNo": {
						"type": "number",
						"format": "double"
					},
					"designation": {
						"$ref": "#/components/schemas/Designation"
					},
					"supervisor": {
						"$ref": "#/components/schemas/User"
					},
					"email": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"doj": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"password": {
						"items": {
							"$ref": "#/components/schemas/Password"
						},
						"type": "array"
					},
					"goalHolder": {
						"$ref": "#/components/schemas/GoalHolder"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"goalHolderId": {
						"type": "number",
						"format": "double"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"teams1": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"userGroupMembers": {
						"items": {
							"$ref": "#/components/schemas/UserGroupMember"
						},
						"type": "array"
					},
					"notifications": {
						"items": {
							"$ref": "#/components/schemas/Notification"
						},
						"type": "array"
					},
					"delete": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"name",
					"employeeId",
					"epfNo",
					"designation",
					"email",
					"password",
					"goalHolder",
					"goalsAvailability",
					"teams",
					"teams1",
					"notifications",
					"delete"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Designation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					},
					"user": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"value",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Password": {
				"properties": {
					"password": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"password",
					"id",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoalHolder": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"teamId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"goalSummaries": {
						"items": {
							"$ref": "#/components/schemas/GoalSummery"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"goalsAvailability",
					"goalSummaries"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserGroupMember": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"teamId": {
						"type": "number",
						"format": "double"
					},
					"teamLeads": {
						"$ref": "#/components/schemas/Team"
					},
					"teamLeadsId": {
						"type": "number",
						"format": "double"
					},
					"userGroup": {
						"$ref": "#/components/schemas/UserGroup"
					},
					"userGroupId": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string",
						"enum": [
							"user",
							"teamMember",
							"teamLead"
						]
					}
				},
				"required": [
					"id",
					"user",
					"team",
					"teamLeads",
					"userGroup",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Notification": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"time": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"image",
					"time",
					"url",
					"read",
					"userId",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Responses": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
...
			"SelectValues": {
				"items": {
					"$ref": "#/components/schemas/SelectValue"
				},
				"type": "array"
			},
			"ControllerResponse_SelectValues_": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"body": {
						"$ref": "#/components/schemas/SelectValues"
					},
					"mask": {}
				},
				"required": [
					"statusCode",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ErrorJson": {
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"body": {}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreationParams": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"epfNo": {
						"type": "number",
						"format": "double"
					},
					"designation": {},
					"supervisor": {},
					"email": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"doj": {
						"type": "string"
					},
					"goalsAvailability": {
						"type": "boolean"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"epfNo",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DataTableResponse_User_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"description": "Bearer authentication - give access to the bearer of this token",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "pms-backend",
		"description": "Performance Management System Backend",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Tharindu vindula"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/common/user/{id}/{teamId}/{cycle}": {
			"get": {
				"operationId": "GetSupervisor",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "cycle",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/goals/review/user/{cycleId}": {
			"get": {
				"operationId": "GetReviewUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Responses"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cycleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/goals/review/user/{userId}/{cycleId}": {
			"get": {
				"operationId": "GetReviewUserForHR",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Responses"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "cycleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/goals/timeline/user/{userId}": {
			"get": {
				"operationId": "GetUserTimeLine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/select-values/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ControllerResponse_SelectValues_"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ids",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					}
				]
			}
		},
...
		"/api/v1/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Responses"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 2,
											"email": "jane@doe.com",
											"name": "Jane Doe",
											"phoneNumbers": []
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": "401",
											"message": "JsonWebTokenError: invalid token || No token provided || JWT does not contain required scope.",
											"body": {}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "notFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": "404",
											"message": "error",
											"body": {}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": "500",
											"message": "internal server error",
											"body": {}
										}
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreationParams"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "EditUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"user": {
										"$ref": "#/components/schemas/UserCreationParams"
									}
								},
								"required": [
									"user"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/request-password-reset": {
			"post": {
				"operationId": "RequestPasswordReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreationParams"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/password-reset": {
			"post": {
				"operationId": "ResetPasswordFromToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreationParams"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								},
								"required": [
									"token",
									"password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataTableResponse_User_"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "employeeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "epfNumber",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreationParams"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"user": {
										"$ref": "#/components/schemas/UserCreationParams"
									}
								},
								"required": [
									"password",
									"user"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"201": {
						"description": "Created"
					},
					"500": {
						"description": "internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"code": "500",
											"message": "internal server error",
											"body": {}
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/refresh-token/{userId}": {
			"get": {
				"operationId": "RefreshToken",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/users/password-change/{userId}": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"hr",
							"manger"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"currentPassword": {
										"type": "string"
									}
								},
								"required": [
									"newPassword",
									"currentPassword"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/password-reset/{userId}": {
			"patch": {
				"operationId": "ResetPassword",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"hr"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									}
								},
								"required": [
									"newPassword"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}